+++ src/gui/widgets/qprogressbar.cpp
@@ -447,7 +447,7 @@ QString QProgressBar::text() const
             || (d->value == INT_MIN && d->minimum == INT_MIN))
         return QString();
 
-    int totalSteps = d->maximum - d->minimum;
+    qint64 totalSteps = qint64(d->maximum) - d->minimum;
 
     QString result = d->format;
     result.replace(QLatin1String("%m"), QString::number(totalSteps));
@@ -461,7 +461,7 @@ QString QProgressBar::text() const
         return result;
     }
 
-    int progress = (d->value - d->minimum) * 100 / totalSteps;
+    int progress = (qreal(d->value) - d->minimum) * 100.0 / totalSteps;
     result.replace(QLatin1String("%p"), QString::number(progress));
     return result;
 }
